# packages/http-contract/openapi.json
{
  "components": {
...
    "securitySchemes": { "bearerAuth": { "scheme": "bearer", "type": "http" } }
  },
  "info": { "title": "Preventive Care Literacy API", "version": "1.0.0" },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/administration/school-workspaces": {
      "post": {
        "operationId": "createSchoolWorkspace",
        "parameters": [
          {
            "in": "header",
            "name": "x-prevcare-csrf",
            "required": true,
            "schema": { "enum": ["1"], "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "displayName": {
                    "maxLength": 200,
                    "minLength": 1,
                    "pattern": ".*\\S.*",
                    "type": "string"
                  },
                  "operationId": { "format": "uuid", "type": "string" },
                  "workspaceId": { "format": "uuid", "type": "string" }
                },
                "required": ["operationId", "workspaceId", "displayName"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "operationId": { "format": "uuid", "type": "string" },
                    "outcome": { "enum": ["created"], "type": "string" },
                    "workspaceId": { "format": "uuid", "type": "string" }
                  },
                  "required": ["operationId", "workspaceId", "outcome"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "status": { "type": "integer" },
                    "title": { "type": "string" },
                    "type": { "type": "string" }
                  },
                  "required": ["type", "title", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "status": { "type": "integer" },
                    "title": { "type": "string" },
                    "type": { "type": "string" }
                  },
                  "required": ["type", "title", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "status": { "type": "integer" },
                    "title": { "type": "string" },
                    "type": { "type": "string" }
                  },
                  "required": ["type", "title", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "status": { "type": "integer" },
                    "title": { "type": "string" },
                    "type": { "type": "string" }
                  },
                  "required": ["type", "title", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "status": { "type": "integer" },
                    "title": { "type": "string" },
                    "type": { "type": "string" }
                  },
                  "required": ["type", "title", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "status": { "type": "integer" },
                    "title": { "type": "string" },
                    "type": { "type": "string" }
                  },
                  "required": ["type", "title", "status", "code"],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  }
}